# docs/api-reference/swagger.yaml
openapi: "3.0.3"
info:
  description: Sandbox API
  version: 1.0.0
  title: Sandbox API
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /accounts/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: The name of the sandbox
        schema:
          type: string
    get:
      tags:
        - account
      operationId: getAccount
      summary: Get information about an account
      responses:
        '200':
          description: The account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        '404':
          description: getAccount account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: getAccount unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    UUID:
      type: string
      description: The UUID of the service. Sandboxes are assigned uuid when booked.
      format: uuid
...
    Annotations:
      description: Key / Value map to provide optional information.
      type: object
      additionalProperties:
        type: string
        minLength: 1
      example:
        guid: abcde
        env_type: ocp4-cluster
...
    Account:
      oneOf:
        - $ref: '#/components/schemas/AwsAccount'
      discriminator:
        propertyName: account_type
        mapping:
          aws: '#/components/schemas/AwsAccount'
...
    AwsAccount:
      type: object
      required:
        - name
        - available
        - account_id
        - zone
        - hosted_zone_id
      properties:
        name:
          type: string
          example: sandbox123
        available:
          description: True if the sandbox has no placement associated.
          type: boolean
          default: true
          example: false
        service_uuid:
          $ref: '#/components/schemas/UUID'
        annotations:
          $ref: '#/components/schemas/Annotations'
        account_id:
          type: string
          example: '0123456'
        zone:
          type: string
          example: sandbox123.domain.com
        hosted_zone_id:
          type: string
          example: Z019612345AH4C7KTTZJ0
        to_cleanup:
          type: boolean
          default: false
          example: true
        conan_status:
          type: string
          example: cleanup in progress
        conan_timestamp:
          type: string
          format: date-time
          example: 2023-03-09T07:22:42+00:00
        conan_hostname:
          type: string
          example: conan2.domain.com
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: integer
          example: 502
        message:
          type: string
          example: Unexpected Error
...